+++ .gitignore
@@ -6,4 +6,4 @@ db_messages.DB
 
 .DS_Store
 
-
+.next/
