+++ app/page.tsx
@@ -1,0 +1,11 @@
+"use client"
+import Chat from '../app/chat/page'
+import React from 'react';
+
+  const App: React.FC = () => {
+  return (
+      <Chat/>
+  );
+};
+export default App;
+
