--- index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Chat-Sergio</title>
-    <link rel="stylesheet" href="src/components/Chat/chat.css">
-</head>
-<body>
-    <div id="root"></div>
-    <script type="module" src="src/index.tsx"></script>
-</body>
-</html>
