--- vite.config.js
@@ -1,10 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-
-// https://vitejs.dev/config/
-export default defineConfig({
-    plugins: [react()],
-    server: {
-        open: true, // Abre el navegador automáticamente
-    },
-});
